+++ packages/networks/src/networks/mainnet.ts
@@ -7,6 +7,7 @@ export const mainnet: NetworkConfig = {
   unlockAddress: '0x3d5409CcE1d45233dE1D4eBDEe74b8E004abDD13',
   multisig: '0xa39b44c4AFfbb56b76a1BF1d19Eb93a5DfC2EBA9',
   name: 'Ethereum',
+  chain: 'ethereum',
   blockTime: 8000,
   subgraph: {
     endpoint: 'https://api.thegraph.com/subgraphs/name/unlock-protocol/unlock',
@@ -37,11 +38,23 @@ export const mainnet: NetworkConfig = {
     name: 'Ether',
     symbol: 'Eth',
     decimals: 18,
+    coingecko: 'ethereum',
   },
   startBlock: 7120795,
   description: 'The most popular network',
   isTestNetwork: false,
   teamMultisig: '0xa39b44c4AFfbb56b76a1BF1d19Eb93a5DfC2EBA9',
+  uniswapV3: {
+    subgraph: 'https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v3',
+    factoryAddress: '0x1F98431c8aD98523631AE4a59f267346ea31F984',
+    quoterAddress: '0xb27308f9F90D607463bb33eA1BeBb41C27CE5AB6',
+  },
+  wrappedNativeCurrency: {
+    name: 'Wrapped Ether',
+    symbol: 'WETH',
+    decimals: 18,
+    address: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',
+  },
   tokens: [
     {
       name: 'Wrapped Ether',
@@ -56,7 +69,7 @@ export const mainnet: NetworkConfig = {
       decimals: 18,
     },
     {
-      name: 'USDCoin',
+      name: 'USD Coin',
       address: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
       symbol: 'USDC',
       decimals: 6,