+++ subgraph/src/unlock.ts
@@ -4,7 +4,7 @@ import { log, BigInt, ethereum } from '@graphprotocol/graph-ts'
 import { NewLock, LockUpgraded } from '../generated/Unlock/Unlock'
 import { PublicLock as PublicLockMerged } from '../generated/templates/PublicLock/PublicLock'
 import { PublicLock } from '../generated/templates'
-import { Lock } from '../generated/schema'
+import { Lock, LockStats, UnlockDailyData } from '../generated/schema'
 import { LOCK_MANAGER } from './helpers'
 
 const ROLE_GRANTED =
@@ -16,6 +16,37 @@ export function handleNewLock(event: NewLock): void {
   // create new lock
   let lockID = lockAddress.toHexString()
   const lock = new Lock(lockID)
+
+  // create new lockStats if not existing
+  let lockStats = LockStats.load('Unlock')
+  if (lockStats === null) {
+    lockStats = new LockStats('Unlock')
+    lockStats.totalLocksDeployed = BigInt.fromI32(0)
+    lockStats.totalKeysSold = BigInt.fromI32(0)
+    lockStats.save()
+  } else {
+    lockStats.totalLocksDeployed = lockStats.totalLocksDeployed.plus(
+      BigInt.fromI32(1)
+    )
+    lockStats.save()
+  }
+
+  // create lockDayData
+  let timestamp = event.block.timestamp.toI32()
+  let dayID = timestamp / 86400
+  let unlockDailyData = UnlockDailyData.load(dayID.toString())
+  if (unlockDailyData === null) {
+    unlockDailyData = new UnlockDailyData(dayID.toString())
+    unlockDailyData.lockDeployed = BigInt.fromI32(1)
+    unlockDailyData.keysSold = BigInt.fromI32(0)
+    unlockDailyData.activeLocks = []
+    unlockDailyData.save()
+  } else {
+    unlockDailyData.lockDeployed = unlockDailyData.lockDeployed.plus(
+      BigInt.fromI32(1)
+    )
+    unlockDailyData.save()
+  }
 
   // fetch lock version
   let lockContract = PublicLockMerged.bind(lockAddress)
@@ -43,12 +74,13 @@ export function handleNewLock(event: NewLock): void {
     lock.symbol = 'KEY'
   }
 
-  let maxKeysPerAddress = lockContract.try_maxKeysPerAddress()
-  if (!maxKeysPerAddress.reverted) {
-    lock.maxKeysPerAddress = maxKeysPerAddress.value
-  } else {
-    // set to 1 when using address instead of tokenId prior to lock v10
-    lock.maxKeysPerAddress = BigInt.fromI32(1)
+  // maxKeysPerAddress set to 1 prior to lock v10
+  lock.maxKeysPerAddress = BigInt.fromI32(1)
+  if (version.ge(BigInt.fromI32(9))) {
+    let maxKeysPerAddress = lockContract.try_maxKeysPerAddress()
+    if (!maxKeysPerAddress.reverted) {
+      lock.maxKeysPerAddress = maxKeysPerAddress.value
+    }
   }
 
   let maxNumberOfKeys = lockContract.try_maxNumberOfKeys()
@@ -61,8 +93,14 @@ export function handleNewLock(event: NewLock): void {
   lock.version = version
   lock.createdAtBlock = event.block.number
 
-  // lock managers are parsed from RoleGranted events
-  lock.lockManagers = []
+  if (version.le(BigInt.fromI32(8))) {
+    // prior to v8, add default lock manager 
+    lock.lockManagers = [event.params.lockOwner]
+  } else {
+    // after v8, lock managers are parsed from `RoleGranted` events
+    lock.lockManagers = []
+  }
+  
 
   lock.save()